Only output warning in verbose mode
If a user has an invalid configuration option in their local pep8 config
file, a warning is emitted to STDOUT causing Code Climate to error the
build when it tries to parse the output as an issue.

For now only output this warning when pep8 is running in verbose mode.
Eventually this should likely output a warning item when using the Code
Climate formatter.